testing/web-platform/tests/atomics/window-and-any-worker/wait/nan-timeout/nan-timeout.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
  <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

// Copyright (C) 2015 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wait does not time out with a NaN timeout
---*/

window.addEventListener("load", function() {
  // Window->Worker.
  async_test(function(t) {
    let worker = new Worker('nan-timeout.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data[0], "ok");
    });
    let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    worker.postMessage([ia]);
    window.setTimeout(() => Atomics.wake(ia, 0), 500);
  }, "atomics-dedicated-worker-wait-nan-timeout");
  // Window->Worker->Worker.
  async_test(function(t) {
    let worker = new Worker('nan-timeout-worker.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data[0], "ok");
    });
    worker.postMessage([]);
  }, "atomics-dedicated-worker-wait-nan-timeout-worker");
  // Window->Worker->SharedWorker.
  async_test(function(t) {
    let worker = new SharedWorker('nan-timeout-shared-worker.js');
    worker.port.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data[0], "ok");
    });
    worker.port.start();
    worker.port.postMessage([]);
  }, "atomics-dedicated-worker-wait-nan-timeout-shared-worker");
  // Window->IFrame->Worker.
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let that = this;
    contentWindow.onmessage = function(e) {
      let worker = new Worker('nan-timeout.js');
      worker.onmessage = that.step_func_done(function(e) {
        assert_equals(e.data[0], "ok");
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia]);
      setTimeout(() => Atomics.wake(ia, 0), 500);
    }
    contentWindow.postMessage([], '*');
  }, "atomics-dedicated-worker-wait-nan-timeout-from-iframe");
});

</script>
